requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.10.0
pandas>=2.0.0
python-docx>=0.8.11
spacy>=3.5.0
beautifulsoup4>=4.12.0
openai>=1.0.0
requests>=2.28.0
openpyxl>=3.1.0
flask>=2.3.0
python-dotenv>=1.0.0 